requirements.txt
Flask==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0 